// api/models/ApiResultCursorPageBaseResponseChatMessageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CursorPageBaseResponseChatMessageResponse } from './CursorPageBaseResponseChatMessageResponse';
export type ApiResultCursorPageBaseResponseChatMessageResponse = {
    success: boolean;
    errCode?: number;
    errMsg?: string;
    data?: CursorPageBaseResponseChatMessageResponse;
};